fix operation config save

+++ src/pages/Banners/List.vue
@@ -4,16 +4,30 @@
       <h2>Banners</h2>
       <div class="box-actions">
         <div class="input-post-icon search">
-          <input type="text" v-model="searchQuery" placeholder="Digite aqui o que deseja encontrar" />
+          <input
+            type="text"
+            v-model="searchQuery"
+            placeholder="Digite aqui o que deseja encontrar"
+          />
           <i v-if="searchQuery === ''" class="icon-icon-search"></i>
-          <i v-else class="bt-clear-search icon-icon-times c-red" @click="searchQuery = ''"></i>
+          <i
+            v-else
+            class="bt-clear-search icon-icon-times c-red"
+            @click="searchQuery = ''"
+          ></i>
         </div>
         <div class="filter" :class="{ active: showFilters }">
-          <a class="bt bt-square bg-white-2 c-light-blue" @click="showFilters = !showFilters">
+          <a
+            class="bt bt-square bg-white-2 c-light-blue"
+            @click="showFilters = !showFilters"
+          >
             <i class="icon-icon-filter"></i>
           </a>
         </div>
-        <base-link to="/banners/new" class="bt bt-square bg-white-2 c-light-blue">
+        <base-link
+          to="/banners/new"
+          class="bt bt-square bg-white-2 c-light-blue"
+        >
           <i class="icon-icon-plus"></i>
         </base-link>
       </div>
@@ -62,7 +76,12 @@
           <tr v-for="item in tableData" :key="item.id">
             <td>
               <div class="img-holder">
-                <img :src="item.image" :alt="item.name" width="96" height="40" />
+                <img
+                  :src="item.image"
+                  :alt="item.name"
+                  width="96"
+                  height="40"
+                />
               </div>
             </td>
             <td>
@@ -114,7 +133,12 @@
                 >
                   <i class="icon-icon-edit"></i>
                 </button>
-                <button @click="handleDelete(item)" type="button" title="apagar" class="bt c-red">
+                <button
+                  @click="handleDelete(item)"
+                  type="button"
+                  title="apagar"
+                  class="bt c-red"
+                >
                   <i class="icon-icon-delete"></i>
                 </button>
               </div>
@@ -133,7 +157,11 @@
         @update-per-page="changePerPage"
       ></pagination>
     </div>
-    <delete-modal @confirmDelete="confirmDelete" :itemName="modal.itemName" :show="modal.visible"></delete-modal>
+    <delete-modal
+      @confirmDelete="confirmDelete"
+      :itemName="modal.itemName"
+      :show="modal.visible"
+    ></delete-modal>
   </div>
 </template>
 <script>
@@ -227,31 +255,47 @@ export default {
         });
       }
     },
-    validateModal() {
+    handleDelete(item) {
+      this.modal.model = item;
+      this.modal.itemName = item.name;
+      this.modal.visible = true;
+    },
+    confirmDelete(val) {
       const self = this;
-      this.$validator.validateAll().then(isValid => {
-        if (isValid) {
-          self.modal.formLoading = true;
-          bannerService.delete(self.modal.model.id).then(
-            response => {
-              self.$notify({
-                type: 'success',
-                message: response.message
-              });
-              self.resetModal();
-              self.pagination.currentPage = 1;
-              self.fetchData();
-            },
-            err => {
-              self.$notify({
-                type: 'danger',
-                message: err.message
-              });
-              self.modal.formLoading = false;
-            }
-          );
-        }
-      });
+      if (val) {
+        this.$validator.validateAll().then(isValid => {
+          if (isValid) {
+            self.modal.formLoading = true;
+            bannerService.delete(self.modal.model.id).then(
+              response => {
+                self.$notify({
+                  type: 'success',
+                  message: response.message
+                });
+                self.resetModal();
+                self.pagination.currentPage = 1;
+                self.fetchData();
+              },
+              err => {
+                if (err.response.status === 400 && err.response.data.message) {
+                  self.$notify({
+                    type: 'warning',
+                    message: err.response.data.message
+                  });
+                } else {
+                  self.$notify({
+                    type: 'danger',
+                    message: err.message
+                  });
+                }
+                self.modal.formLoading = false;
+              }
+            );
+          }
+        });
+      } else {
+        self.resetModal();
+      }
     }
   },
   watch: {
